renamed file and updated with pr comments
