+++ main.go
@@ -10,7 +10,7 @@ import (
 	"go.viam.com/rdk/module"
 	"go.viam.com/utils"
 
-	"github.com/viam-labs/ocean-prefilter/ocean_prefilter"
+	"github.com/viam-labs/ocean-prefilter/oceanprefilter"
 )
 
 func main() {
@@ -27,7 +27,7 @@ func mainWithArgs(ctx context.Context, args []string, logger logging.Logger) (er
 
 	// Models and APIs add helpers to the registry during their init().
 	// They can then be added to the module here.
-	err = myMod.AddModelFromRegistry(ctx, vision.API, ocean_prefilter.Model)
+	err = myMod.AddModelFromRegistry(ctx, vision.API, oceanprefilter.Model)
 	if err != nil {
 		return err
 	}
